train/requirements.txt
numpy
pandas
gcsfs
google-cloud-storage
kubeflow-fairing
grpcio

